+++ internal/acceptance_tests/blocks/fastly_user_service_authorization.tf
@@ -1,4 +1,4 @@
-resource "fastly_service_authorization" "test" {
+resource "fastly_user_service_authorization" "test" {
   service_id = {{.SERVICE_ID_REF}}
   user_id    = "{{.USER_ID}}"
   permission = "{{.PERMISSION}}"