+++ internal/provider/provider.go
@@ -126,7 +126,6 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/productenablement"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/resourcelink"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/secretstore"
-	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/serviceauthorization"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/servicecdn"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/servicecdnauto"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/servicecompute"
@@ -141,6 +140,7 @@ import (
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/tlssubscription"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/tlssubscriptionvalidation"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/tsigkey"
+	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/userserviceauthorization"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/resources/vcl"
 	"github.com/fastly/terraform-provider-fastly-beta/internal/version"
 )
@@ -286,7 +286,6 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		productenablement.NewNGWAFResource,
 		resourcelink.NewResource,
 		secretstore.NewResource,
-		serviceauthorization.NewResource,
 		servicecdn.NewResource,
 		servicecdnauto.NewResource,
 		servicecompute.NewResource,
@@ -300,6 +299,7 @@ func (p *fastlyProvider) Resources(_ context.Context) []func() resource.Resource
 		tlssubscription.NewResource,
 		tlssubscriptionvalidation.NewResource,
 		tsigkey.NewResource,
+		userserviceauthorization.NewResource,
 	}
 }